# art_admin_backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateMeditationContentRequest:
    properties:
      audio_url:
        type: string
      category:
        type: string
      cover_image:
        type: string
      description:
        type: string
      difficulty_level:
        maximum: 5
        minimum: 1
        type: integer
      duration:
        minimum: 1
        type: integer
      is_public:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        type: string
    required:
    - audio_url
    - category
    - duration
    - title
    type: object
...
  request.UpdateMeditationContentRequest:
    properties:
      audio_url:
        type: string
      category:
        type: string
      cover_image:
        type: string
      description:
        type: string
      difficulty_level:
        type: integer
      duration:
        type: integer
      id:
        type: integer
      is_public:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - id
    type: object
...
  response.MeditationContentDetail:
    properties:
      audio_url:
        type: string
      category:
        type: string
      category_label:
        type: string
      cover_image:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      difficulty_level:
        type: integer
      duration:
        type: integer
      id:
        type: integer
      is_public:
        type: boolean
      like_count:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  response.MeditationContentItem:
    properties:
      audio_url:
        type: string
      category:
        type: string
      category_label:
        type: string
      cover_image:
        type: string
      difficulty_level:
        type: integer
      duration:
        type: integer
      id:
        type: integer
      is_public:
        type: boolean
      like_count:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      view_count:
        type: integer
    type: object
...
  response.PaginatedData:
    properties:
      current:
        description: 当前页码
        type: integer
      records:
        description: 记录列表
      size:
        description: 每页条数
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
    type: object
...
host: localhost:48080
info:
  contact: {}
  description: Art Admin Backend API Documentation
  title: Art Admin API
  version: "1.0"
paths:
...
  /api/meditation-content/{id}:
    get:
      consumes:
      - application/json
      description: 获取冥想内容详细信息
      parameters:
      - description: 冥想内容ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MeditationContentDetail'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取冥想内容详情
      tags:
      - 冥想管理
  /api/meditation-content/categories:
    get:
      consumes:
      - application/json
      description: 获取所有冥想分类
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取冥想分类列表
      tags:
      - 冥想管理
  /api/meditation-content/category/{category}:
    get:
      consumes:
      - application/json
      description: 根据分类获取冥想内容列表
      parameters:
      - description: 冥想分类
        in: path
        name: category
        required: true
        type: string
      - default: 10
        description: 返回数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.MeditationContentItem'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 根据分类获取冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/create:
    post:
      consumes:
      - application/json
      description: 创建新的冥想内容
      parameters:
      - description: 冥想内容信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateMeditationContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除冥想内容
      parameters:
      - description: 冥想内容ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/difficulty-levels:
    get:
      consumes:
      - application/json
      description: 获取所有难度等级
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: integer
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取难度等级列表
      tags:
      - 冥想管理
  /api/meditation-content/difficulty/{difficulty}:
    get:
      consumes:
      - application/json
      description: 根据难度等级获取冥想内容列表
      parameters:
      - description: 难度等级
        in: path
        name: difficulty
        required: true
        type: integer
      - default: 10
        description: 返回数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.MeditationContentItem'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 根据难度等级获取冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/like/{id}:
    post:
      consumes:
      - application/json
      description: 点赞冥想内容
      parameters:
      - description: 冥想内容ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 点赞成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 点赞冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/list:
    get:
      consumes:
      - application/json
      description: 分页查询冥想内容列表
      parameters:
      - description: 当前页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - description: 冥想分类
        in: query
        name: category
        type: string
      - description: 难度等级
        in: query
        name: difficultyLevel
        type: integer
      - description: 是否公开
        in: query
        name: isPublic
        type: boolean
      - description: 最小时长(秒)
        in: query
        name: minDuration
        type: integer
      - description: 最大时长(秒)
        in: query
        name: maxDuration
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取冥想内容列表
      tags:
      - 冥想管理
  /api/meditation-content/popular:
    get:
      consumes:
      - application/json
      description: 获取热门冥想内容列表
      parameters:
      - default: 10
        description: 返回数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.MeditationContentItem'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取热门冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/search:
    get:
      consumes:
      - application/json
      description: 搜索冥想内容
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        required: true
        type: string
      - description: 当前页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 搜索冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/unlike/{id}:
    post:
      consumes:
      - application/json
      description: 取消点赞冥想内容
      parameters:
      - description: 冥想内容ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 取消点赞成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 取消点赞冥想内容
      tags:
      - 冥想管理
  /api/meditation-content/update:
    put:
      consumes:
      - application/json
      description: 更新冥想内容信息
      parameters:
      - description: 冥想内容信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateMeditationContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新冥想内容
      tags:
      - 冥想管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"